In development

+++ uninstall.sh
@@ -8,16 +8,16 @@ systemctl stop vpnconfig_ovpn
 systemctl disable vpnconfig_pihole
 systemctl stop vpnconfig_pihole
 
-
 rm /etc/systemd/system/vpnconfig_ovpn.service
 rm /etc/systemd/system/vpnconfig_pihole.service
 systemctl daemon-reload
-
 
 docker network rm $NETWORK
 
 rm /usr/local/bin/ovpn
 rm /usr/local/bin/pihole
 
-rm -r $VOLUMES_PATH/ovpn
-rm -r $VOLUMES_PATH/pihole
+if [ "$1" = "--with-data" ]; then
+  rm -r $VOLUMES_PATH/ovpn
+  rm -r $VOLUMES_PATH/pihole
+fi
